requirements.txt
# Core dependencies
google-generativeai>=0.4.1
chromadb>=0.4.22
langchain>=0.1.9
langchain-google-genai>=0.0.11
pypdf>=4.0.1
python-dotenv>=1.0.1

# Document processing
markdown>=3.5.2

# Testing
pytest>=8.0.2
pytest-cov>=4.1.0

# Code quality
black>=24.2.0
flake8>=7.0.0
isort>=5.13.2

# CLI
click>=8.1.7
rich>=13.7.0 